requirements.txt
# --- Data Processing / Analytics ---
numpy
pandas
pyspark
SQLAlchemy
psycopg2-binary

# --- AWS / Cloud ---
boto3
botocore
s3transfer
jmespath

# --- Kafka / Streaming ---
kafka-python
py4j

# --- Logging / JSON ---
python-json-logger

# --- Date/Time Utilities ---
python-dateutil
pytz
tzdata

# --- Interactive / Shell / Notebook ---
ipython
ipython_pygments_lexers
matplotlib-inline
prompt_toolkit
pexpect
ptyprocess
wcwidth

# --- Linting / Code Quality ---
flake8
pycodestyle
pyflakes
mccabe

# --- Syntax Highlighting / Parsing ---
Pygments
parso
jedi

# --- JSON / Schema / Validation ---
jsonschema
jsonschema-specifications
referencing
rpds-py

# --- Utilities / Misc ---
asttokens
attrs
decorator
executing
pure_eval
six
stack-data
traitlets
typing_extensions
python-dotenv
PyYAML
pydantic
pydantic-settings